Removed upload app from project.
Its functionality is being folded into the datapoint app. Ultimately all this app existed to do can be replaced with a couple of templates and one Class based view in the datapoint app. This will simplify the structure of the app and hopefully make more sense for anyone else who approached the code.