Implemented REST API for annotation according to Two Scoops of Django best practices.
This means that all API calls will be to `lackawanna.com/api/...` and that they are very quick to create basic ones. I foresee the need for lots and lots of these calls being available but these annotation ones are absolutely key.

The next things to improve with the API are:
* Authentication and permission
* Logging

Neither are 100% necessary for now so they will be dealt with later.