Removed taggit from the system.
Going for a more basic and model version of a tagging system. Taggit, while working well in the past, was not compatible with the Django Rest Framework REST implementation. This is a blocker and even if there will be some features missing, these can hopefully be replaced if needed. For now, this will do. Next stage is to refactor annotations into the datapoint application.