Removed Angular from the project.
Based on all the changes and 2.0 looming, building on 1.3 is insane. A lot of what Angular does could be useful but it mixed front and backend too much for me. No separation of concerns. It was also another tool to learn when on this project's scale I could recreate what I need in JavaScript.