Moved to working on dissertation with Overleaf.
Overleaf (aka WriteLatex) offers a link into their system through git, meaning I can work on it on my local machine as well as on their online thing. This is easier than having as a branch within this repo. I can share the link with others to get feedback on my writing while still being able to edit locally. This is now a dead branch effectively.