Added exception to deal with serialisation errors.
Because the way that we retrieve the ranges array from validated_data has to be clunky, it doesn't react well if it isn't there and a PATCH call is made. Have placed the work done to it in a try catch with a logging if something goes wrong. Kind of calls for the need to log everything.