Refactored to more generalised large_file nomenclature.
Built support in for large_files in datapoint_detail for:
* video
* audio
* pdf

Currently I would imagine all other file types to be able to use the existing pass-through upload method which allows for pre-processing.

Eventually all files would be directly uploaded to storage and then an asynchronous task queue could complete processing tasks in the background on a different thread. FUTURE TO DO