Add script to automatically run database migrations.
This was leading to the database becoming out of sync on local
environments and in deployment, leading to inconsistencies. There
was a bug from a migration that hadn't been applied that was causing
issues in the file upload of .docx files as the extension was too long for
the original limit of 5 chars.